requirements.txt
beautifulsoup4>=4.2.1
django-braces>=1.0.0
django_celery>=3.1.10
django-crispy-forms>=1.2.3
djangorestframework>=2.3.12
feedparser>=5.1.2
django-haystack==2.1.0
httplib2>=0.7.5
requests>=2.2.1
simplejson>=2.6.2
twitter>=1.9.2
Whoosh==2.6.0
pyaml==14.05.6
pytz==2014.2

# Testing
celery>=3.1.10
MySQL-python>=1.2.5
selenium==2.42.0
